tests/Neba.Infrastructure.UnitTests/Middleware/ExceptionHandlingMiddlewareTests.cs
using System.Net;
using Microsoft.AspNetCore.Http;
using Microsoft.Extensions.Logging;
using Neba.Infrastructure.Middleware;

namespace Neba.Infrastructure.UnitTests.Middleware;

public sealed class ExceptionHandlingMiddlewareTests
{
    private readonly ILogger<ExceptionHandlingMiddleware> _logger;
    private readonly ExceptionHandlingMiddleware _middleware;
    private readonly RequestDelegate _next;

    public ExceptionHandlingMiddlewareTests()
    {
        _next = Substitute.For<RequestDelegate>();
        _logger = Substitute.For<ILogger<ExceptionHandlingMiddleware>>();
        _middleware = new ExceptionHandlingMiddleware(_next, _logger);
    }

    [Fact]
    public async Task InvokeAsync_WhenNextThrowsException_ShouldLogExceptionAndSetResponse()
    {
        // Arrange
        var context = new DefaultHttpContext
        {
            Response =
            {
                Body = new MemoryStream()
            }
        };
        var exceptionMessage = "Test exception";
        _next.When(n => n(context)).Do(callInfo => throw new Exception(exceptionMessage));

        // Act
        await _middleware.InvokeAsync(context);

        // Assert
        context.Response.StatusCode.Should().Be((int)HttpStatusCode.InternalServerError);
        context.Response.Body.Seek(0, SeekOrigin.Begin);
        var reader = new StreamReader(context.Response.Body);
        var responseBody = await reader.ReadToEndAsync();
        responseBody.Should().Contain("An unexpected error occurred");
    }

    [Fact]
    public async Task InvokeAsync_WhenNextDoesNotThrowException_ShouldNotSetResponse()
    {
        // Arrange
        var context = new DefaultHttpContext
        {
            Response =
            {
                Body = new MemoryStream()
            }
        };

        // Act
        await _middleware.InvokeAsync(context);

        // Assert
        context.Response.StatusCode.Should().Be((int)HttpStatusCode.OK);
        context.Response.Body.Seek(0, SeekOrigin.Begin);
        var reader = new StreamReader(context.Response.Body);
        var responseBody = await reader.ReadToEndAsync();
        responseBody.Should().BeEmpty();
    }
}